Started the Home Page of the website
I added the homepage of the website, which is where users get
redirected to when they click enter on the landing page. I’ve put some
buttons representing the different categories. I’ve applied some basic
CSS styles to them.

Next is too perhaps use a picture logo instead of an h1 element so it
can resize properly. Also time for implementing the pictures in the
gallery.