Website Street Images - Updated
Massive update.

I’m really happy I got my images to properly resize. It was very
difficult since each image doesn’t have the same aspect ratio. However
each image has the same width of 500px. I worked with that and the fact
that all images height would be based on the first image height. I
wrote a javascript function to calculate a new height for each wrapper
div that is based on the auto height of the first image. This function
is called before load time and when window resizes. It looks to be a
little bit slow, not that much. But its the best solution I’ve found so
far. Object fit is key to not distorting the pictures.

Next: To change h1 to a jpg so that I can resize it dynamically. Also
adjust navigation links on first css breakpoint. Lastly, fill the other
pages with images.