Big Changes - Content Filter Works!
I have made multiple media queries and for specific breakpoints, higher
resolution pictures are put into place.

One of my hardest challenged when I was working on my website last
summer was to implement a content filtering system into a html page
that was already very branched out. The modifications were too vast and
complicated so I just gave up.

This time, my html structure was still simple enough to be able to
implement a content filtering script. All I had to do was first
identify when a user clicks on any of the navigation items (Street,
land, faces…), then get the respective class name and apply a selected
class to all images that have that class name. Elements who do not have
the selected class are set to display none.

Because eventually it would be too much to load all images at once, for
all other tabs I put the image source as a data-src attribute and later
loaded the img src attribute when a user clicked on that tab.

Next steps is too setup the next and previous buttons to allow user to
navigate pictures in the full mode and also allow users to press
keyboard arrows.