Fixed Bugs - Image On Load
Fixed a lot of bugs regarding image on load event. A lot of code was dependent on the image being loaded and I was just using a SetTimeOut function which is just naive because an image can take various amounts of time to load depending on network. Since I was testing on my home computer images always loaded super fast.

Problems came when I tried testing on the server. So I fixed all of that by creating a new Image using vanilla js and settings its source to the first image and attaching an onload event. Now it properly waits until the first image is loaded before even calling any function.